requirements_full.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-cors==0.8.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.71.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
apache-tvm-ffi==0.1.6
astor==0.8.1
attrs==25.4.0
blake3==1.0.8
cachetools==6.2.4
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorful==0.5.8
compressed-tensors==0.12.2
cuda-bindings==13.1.1
cuda-pathfinder==1.3.3
cuda-python==13.1.1
datasets==4.2.0
depyf==0.20.0
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.1
email-validator==2.3.0
fastapi==0.127.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
filelock==3.20.1
flash_attn @ file:///opt/transfer/flash_attn-2.7.4.post1%2B25.12-40345937-cp312-cp312-linux_aarch64.whl#sha256=7db5f9e748949c02e3ff65869873fea69269770486bad9dc650a07a1e83387f2
flashinfer-cubin @ file:///opt/transfer/flashinfer_cubin-0.6.0rc2%2B9c4af0a1.nvinternal.25.12.post1.devel.cu131-40598484-py3-none-any.whl#sha256=9efcb7804078ab6e2e324e6a70331481cf58c779a91ce58df71e2484a60792a9
flashinfer-python @ file:///opt/transfer/flashinfer_python-0.6.0rc2%2B9c4af0a1.nvinternal.25.12.post1.devel.cu131-40598484-py3-none-any.whl#sha256=4c9c515a0b99a7858d5deaad8e0dcffcf7c1add865504187e660b035e1a162bc
frozenlist==1.8.0
fsspec==2025.9.0
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.28.1
google-auth==2.45.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
hf_transfer==0.1.9
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.1
interegular==0.3.3
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lark==1.2.2
llguidance==1.3.0
llvmlite==0.46.0
lm-format-enforcer==0.11.3
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mistral_common==1.8.8
model-hosting-container-standards==0.1.12
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.6.1
ninja==1.13.0
numba==0.63.1
numpy==2.1.0
nvidia-cudnn-frontend==1.17.0
nvidia-cutlass-dsl==4.3.4
nvidia-ml-py==13.590.44
omegaconf==2.3.0
openai==2.14.0
openai-harmony==0.0.8
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
outlines_core==0.2.11
packaging==25.0
pandas==3.0.1
partial-json-parser==0.2.1.1.post7
peft==0.18.1
pillow==12.0.0
platformdirs==4.5.1
-e git+ssh://git@github.com/swiss-ai/posttraining.git@2d48b4519ad1b59b0e6ccb23cf79a05f3e5905c6#egg=posttraining
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.1.3
py-cpuinfo==9.0.0
py-spy==0.4.1
pyarrow==23.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.53.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rewarduq==0.1.0
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scipy==1.16.3
sentencepiece==0.2.1
sentry-sdk==2.48.0
setproctitle==1.3.7
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
smart_open==7.5.0
smmap==5.0.3
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
starlette==0.50.0
supervisor==4.3.0
sympy==1.14.0
tabulate==0.9.0
tiktoken==0.12.0
tokenizers==0.22.1
torch @ file:///opt/transfer/torch-2.10.0a0%2Bb4e4ee81d3.nv25.12-40345937-cp312-cp312-linux_aarch64.whl#sha256=aae06bbfe23a557da17ecfcff41ff2a9a8d2d61363899eac777771f3ccfb273e
torchvision @ file:///opt/transfer/torchvision-0.25.0a0%2Bca221243-40345937-cp312-cp312-manylinux_2_28_aarch64.whl#sha256=021cbd5bcf3a47908b2ac29a648ca02046824a3564f41f76efeff7fb3aa10af6
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.1
triton @ file:///tmp/dist/pytorch_triton-3.5.1%2Bgitcced7793-cp312-cp312-linux_aarch64.whl#sha256=bb20264a42d6f01ae5d3407e08b13728e49d3dfc45ed39afc831d821f62dcfb8
trl==0.23.0
typer==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
uvicorn==0.40.0
uvloop==0.22.1
virtualenv==20.35.4
vllm @ file:///opt/transfer/vllm-0.12.0%2B35a9f223.nv25.12.post1.cu131-40694329-cp38-abi3-linux_aarch64.whl#sha256=851a058c2b27f379e77114eeea03651cf5be177f29730f554b0ba4836c62f5bc
wandb==0.23.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.0.1
xformers @ file:///opt/transfer/xformers-0.0.33%2Bnv25.12.post1-40694329-cp39-abi3-linux_aarch64.whl#sha256=b0ef18dcea7cf639d685f564f0aa237c3eb093cbe06e90a19ec2136ae48976f2
xgrammar==0.1.27
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
